+++ geoget/version.py
@@ -1,0 +1,2 @@
+version = '0.0.1.dev-42351e0'
+short_version = '0.0.1'
